+++ test/test_stateful.py
@@ -196,7 +196,7 @@ def test_misspelled_parameter(schema_url, parameter, message):
     case = schema["/users/{user_id}"]["GET"].make_case()
     link = schema["/users/"]["POST"].links["201"]["#/paths/~1users~1{user_id}/get"]
     with pytest.raises(ValueError, match=re.escape(message)):
-        link.set_data(case, context=expressions.ExpressionContext(case=case, response=None))
+        link.set_data(case, elapsed=1.0, context=expressions.ExpressionContext(case=case, response=None))
 
 
 @pytest.mark.parametrize(